refactoring : moved everything from 'doclet/expander' to 'doclet'
