+++ sporniket-javabeans-doclet/src/main/java/com/sporniket/libre/javabeans/doclet/BuilderGenerator.java
@@ -1,4 +1,4 @@
-package com.sporniket.libre.javabeans.doclet.expander;
+package com.sporniket.libre.javabeans.doclet;
 
 public interface BuilderGenerator extends JavaSourceGenerator
 {
