+++ sporniket-javabeans-doclet/src/main/java/com/sporniket/libre/javabeans/doclet/DistillerDoclet.java
@@ -1,10 +1,10 @@
 /**
  *
  */
-package com.sporniket.libre.javabeans.doclet.expander;
+package com.sporniket.libre.javabeans.doclet;
 
-import static com.sporniket.libre.javabeans.doclet.expander.UtilsClassDoc.updateKnowClasses;
-import static com.sporniket.libre.javabeans.doclet.expander.UtilsClassname.*;
+import static com.sporniket.libre.javabeans.doclet.UtilsClassDoc.updateKnowClasses;
+import static com.sporniket.libre.javabeans.doclet.UtilsClassname.*;
 import static java.util.Arrays.asList;
 import static java.util.stream.Collectors.toCollection;
 
@@ -19,10 +19,10 @@ import java.util.Map;
 import java.util.Set;
 import java.util.TreeSet;
 
-import com.sporniket.libre.javabeans.doclet.expander.basic.BasicBuilderGenerator;
-import com.sporniket.libre.javabeans.doclet.expander.basic.BasicJavabeanGenerator;
-import com.sporniket.libre.javabeans.doclet.expander.basic.BasicRawPojoGenerator;
-import com.sporniket.libre.javabeans.doclet.expander.basic.Builder;
+import com.sporniket.libre.javabeans.doclet.basic.BasicBuilderGenerator;
+import com.sporniket.libre.javabeans.doclet.basic.BasicJavabeanGenerator;
+import com.sporniket.libre.javabeans.doclet.basic.BasicRawPojoGenerator;
+import com.sporniket.libre.javabeans.doclet.basic.Builder;
 import com.sun.javadoc.ClassDoc;
 import com.sun.javadoc.LanguageVersion;
 import com.sun.javadoc.PackageDoc;
