+++ sporniket-javabeans-doclet/src/main/java/com/sporniket/libre/javabeans/doclet/JavabeanGenerator.java
@@ -1,7 +1,7 @@
 /**
  *
  */
-package com.sporniket.libre.javabeans.doclet.expander;
+package com.sporniket.libre.javabeans.doclet;
 
 /**
  * @author dsporn
