+++ sporniket-javabeans-doclet/src/main/java/com/sporniket/libre/javabeans/doclet/basic/package-info.java
@@ -3,4 +3,4 @@
  * @author dsporn
  *
  */
-package com.sporniket.libre.javabeans.doclet.expander.basic;
+package com.sporniket.libre.javabeans.doclet.basic;
