+++ sporniket-javabeans-doclet/src/main/java/com/sporniket/libre/javabeans/doclet/package-info.java
@@ -4,4 +4,4 @@
  * @author dsporn
  *
  */
-package com.sporniket.libre.javabeans.doclet.expander;
+package com.sporniket.libre.javabeans.doclet;
